+++ src/parsers/problem/LibraryCheckerProblemParser.ts
@@ -1,0 +1,30 @@
+import { Sendable } from '../../models/Sendable';
+import { TaskBuilder } from '../../models/TaskBuilder';
+import { htmlToElement } from '../../utils/dom';
+import { Parser } from '../Parser';
+
+export class LibraryCheckerProblemParser extends Parser {
+  public getMatchPatterns(): string[] {
+    return ['https://judge.yosupo.jp/problem/*'];
+  }
+
+  public async parse(url: string, html: string): Promise<Sendable> {
+    const elem = htmlToElement(html);
+    const task = new TaskBuilder('Library Checker').setUrl(url);
+
+    task.setName(elem.querySelector('.uk-container > h1').textContent);
+
+    task.setTimeLimit(5000);
+    task.setMemoryLimit(1024);
+
+    const preBlocks = [...elem.querySelectorAll('pre')].filter(block => block.querySelector('code') === null);
+    for (let i = 0; i < preBlocks.length - 1; i += 2) {
+      const input = preBlocks[i].textContent;
+      const output = preBlocks[i + 1].textContent;
+
+      task.addTest(input, output);
+    }
+
+    return task.build();
+  }
+}